reminders improved and dynamically added as buttons to dashboard

+++ Home2/app/src/main/java/com/example/home/ReminderController.java
@@ -50,7 +50,7 @@ public class ReminderController {
 
 
 
-    public static void deleteReminder(long reminderID) {
+    public static void deleteReminder(int reminderID) {
         Iterator<Reminder> iterator = Values.RemindersList.iterator();
         while (iterator.hasNext()) {
             Reminder reminder = iterator.next();
